Updated templates of device sdk in go
