--- examples/command_device/main.go
@@ -1,66 +1,0 @@
-package main
-
-import (
-	"context"
-	"errors"
-	"flag"
-	"io"
-	"os"
-	"syscall"
-	"time"
-
-	command_devicev1 "github.com/maxthom/mir/examples/command_device/gen/command_device/v1"
-	"github.com/maxthom/mir/internal/libs/boiler/mir_signals"
-	"github.com/maxthom/mir/pkgs/device/mir"
-	"google.golang.org/protobuf/reflect/protoreflect"
-)
-
-func main() {
-	ctx, cancel := context.WithCancel(context.Background())
-	mir_signals.Notify(syscall.SIGHUP, syscall.SIGINT, syscall.SIGQUIT)
-
-	deviceId := flag.String("id", "0xf86cmd", "Device ID")
-	flag.Parse()
-
-	m, err := mir.Builder().
-		DeviceId(*deviceId).
-		Target("nats://127.0.0.1:4222").
-		LogLevel(mir.LogLevelDebug).
-		LogWriters([]io.Writer{os.Stdout}).
-		DefaultConfigFile(mir.Yaml).
-		Schema(
-			command_devicev1.File_command_device_v1_command_proto,
-		).
-		Build()
-	if err != nil {
-		panic(err)
-	}
-	l := m.Logger()
-
-	l.Info().Msg("Mir is ready for launch... Launching!")
-	mirWg, err := m.Launch(ctx)
-	if err != nil {
-		l.Error().Err(err).Msg("Abort launch error")
-		os.Exit(1)
-	}
-	l.Info().Msg("Mir is at maxq and nominal")
-
-	m.HandleCommand(
-		&command_devicev1.ChangePower{},
-		func(m protoreflect.ProtoMessage) (protoreflect.ProtoMessage, error) {
-			cmd := m.(*command_devicev1.ChangePower)
-
-			l.Info().Msg("ChangePower command received " + cmd.String())
-			time.Sleep(2 * time.Second)
-			l.Info().Msg("Power changed successfully")
-
-			//return &command_devicev1.ChangePowerResponse{Success: true}, nil
-			return nil, errors.New("Fuck this shit")
-		},
-	)
-
-	mir_signals.WaitForOsSignals(func() {
-		cancel()
-		mirWg.Wait()
-	})
-}
