--- examples/command_device/proto/command_device/v1/utils.proto
@@ -1,4 +1,0 @@
-syntax = "proto3";
-
-package command_device.v1;
-option go_package = "github.com/maxthom/mir/examples/command_device/gen/command_device/v1";
