--- examples/config_device/buf.gen.yaml
@@ -1,9 +1,0 @@
-version: v2
-managed:
-  enabled: true
-plugins:
-  - local: protoc-gen-go
-    out: examples/config_device/gen
-    opt: paths=source_relative
-inputs:
-  - directory: examples/config_device/proto
