--- examples/config_device/proto/config_device/v1/config.proto
@@ -1,47 +1,0 @@
-syntax = "proto3";
-
-package config_device.v1;
-option go_package = "github.com/maxthom/mir/examples/config_device/gen/config_device/v1";
-
-import "mir/device/v1/mir.proto";
-
-message Config {
-	option (mir.device.v1.message_type) = MESSAGE_TYPE_TELECONFIG;
-	option (mir.device.v1.meta) = {
-		tags: [{
-			key: "building",
-			value: "A"
-		}, {
-			key: "floor",
-			value: "1"
-		}]
-	};
-
-	int32 power = 1;
-	bool valve_open = 2;
-	int32 gaz_level = 3;
-	Args args = 4;
-}
-
-message Args {
- int32 pizza_count = 1;
- repeated string names = 2;
- map<string, int32> ingredients = 3;
-}
-
-message PhoneConfig {
-	option (mir.device.v1.message_type) = MESSAGE_TYPE_TELECONFIG;
-	option (mir.device.v1.meta) = {
-		tags: [{
-			key: "building",
-			value: "A"
-		}, {
-			key: "floor",
-			value: "1"
-		}]
-	};
-
-	int32 country_code = 1;
-	string number = 2;
-	int32 extension = 3;
-}
