--- examples/config_device/proto/config_device/v1/utils.proto
@@ -1,4 +1,0 @@
-syntax = "proto3";
-
-package config_device.v1;
-option go_package = "github.com/maxthom/mir/examples/config_device/gen/config_device/v1";
