--- examples/device_template/schemav1/schema.proto
@@ -1,37 +1,0 @@
-syntax = "proto3";
-
-package schemav1;
-option go_package = "<project_path>/schemav1";
-
-import "mir/device/v1/mir.proto";
-
-message Telemetry {
-	option (mir.device.v1.message_type) = MESSAGE_TYPE_TELEMETRY;
-
-	int64 ts = 1 [(mir.device.v1.timestamp) = TIMESTAMP_TYPE_NANO];
-	int32 temperature = 2;
-	int32 pressure = 3;
-	int32 humidity = 4;
-}
-
-message Command {
-	option (mir.device.v1.message_type) = MESSAGE_TYPE_TELECOMMAND;
-
-	int64 power = 1;
-}
-
-message CommandResponse {
-  bool success = 1;
-}
-
-message DesiredProperty {
-	option (mir.device.v1.message_type) = MESSAGE_TYPE_TELECONFIG;
-
-	int64 datarate_sec = 1;
-}
-
-message ReportedProperty {
-	option (mir.device.v1.message_type) = MESSAGE_TYPE_TELECONFIG;
-
-	int64 datarate_sec = 1;
-}
