--- examples/telemetry_device/buf.gen.yaml
@@ -1,9 +1,0 @@
-version: v2
-managed:
-  enabled: true
-plugins:
-  - local: protoc-gen-go
-    out: examples/telemetry_device/gen
-    opt: paths=source_relative
-inputs:
-  - directory: examples/telemetry_device/proto
