--- examples/telemetry_device/proto/telemetry_device/v1/telemetry.proto
@@ -1,96 +1,0 @@
-syntax = "proto3";
-
-package telemetry_device.v1;
-option go_package = "github.com/maxthom/mir/examples/telemetry_device/gen/telemetry_device/v1";
-
-import "mir/device/v1/mir.proto";
-
-message EnvironmentTlm {
-	option (mir.device.v1.message_type) = MESSAGE_TYPE_TELEMETRY;
-	option (mir.device.v1.meta) = {
-		tags: [{
-			key: "building",
-			value: "A"
-		}, {
-			key: "floor",
-			value: "1"
-		}]
-	};
-
-	mir.device.v1.Timestamp ts = 1 [(mir.device.v1.timestamp) = TIMESTAMP_TYPE_FRACTION];
-	int32 temperature = 2;
-	int32 pressure = 3;
-	int32 humidity = 4;
-	int32 wind_speed = 5;
-}
-
-message PowerConsuption {
-	option (mir.device.v1.message_type) = MESSAGE_TYPE_TELEMETRY;
-	option (mir.device.v1.meta) = {
-		tags: [{
-			key: "building",
-			value: "A"
-		}, {
-			key: "floor",
-			value: "2"
-		}]
-	};
-
-	int64 ts = 1 [(mir.device.v1.timestamp) = TIMESTAMP_TYPE_NANO];
-	double voltage = 2;
-	double amp = 3;
-	double power = 4;
-}
-
-message Constelletion {
-	option (mir.device.v1.message_type) = MESSAGE_TYPE_TELEMETRY;
-
-	int64 ts = 1 [(mir.device.v1.timestamp) = TIMESTAMP_TYPE_SEC];
-	repeated Satellite satellites = 2;
-}
-
-message Satellite {
-    option (mir.device.v1.meta) = {
-		tags: [{
-			key: "space",
-			value: "deep"
-		}, {
-			key: "orbit",
-			value: "geo"
-		}]
-	};
-	string id = 1;
-	int32 signal_strength = 2;
-	Gps gps = 3;
-}
-
-message Gps {
-	double latitude = 1 [(mir.device.v1.field_meta) = {
-	    tags: [{
-		key: "gpsx",
-		value: "x"
-		}]
-	}];
-	double longitude = 2 [(mir.device.v1.field_meta) = {
-	    tags: [{
-			key: "gpsy",
-			value: "y"
-		}]
-	}];
-	double altitude = 3 [(mir.device.v1.field_meta) = {
-	    tags: [{
-			key: "gpsz",
-			value: "z"
-		}]
-	}];
-}
-
-message ChangePower {
-	option (mir.device.v1.message_type) = MESSAGE_TYPE_TELECOMMAND;
-
-	int32 power = 1;
-}
-
-message ChangePowerResponse {
-    bool success = 1;
-}
