--- examples/telemetry_device/proto/telemetry_device/v1/utils.proto
@@ -1,4 +1,0 @@
-syntax = "proto3";
-
-package telemetry_device.v1;
-option go_package = "github.com/maxthom/mir/examples/telemetry_device/gen/telemetry_device/v1";
