+++ examples/templates/go_device_buf/buf.gen.yaml
@@ -1,0 +1,14 @@
+version: v2
+managed:
+  enabled: true
+  # override:
+  #   - file_option: go_package_prefix
+  #     value: <module_path>/proto
+plugins:
+  - local: protoc-gen-go
+    out: proto
+    opt: paths=source_relative
+inputs:
+  - directory: proto
+    exclude_paths:
+      - proto/mir
