+++ examples/templates/go_device_buf/proto/mir/device/v1/mir.proto
