+++ examples/templates/go_device_protoc/proto/mir/device/v1/mir.proto
@@ -1,0 +1,68 @@
+
+syntax = "proto3";
+
+package mir.device.v1;
+option go_package = "github.com/maxthom/mir/pkgs/device/gen/proto/mir/device/v1";
+
+import "google/protobuf/descriptor.proto";
+
+extend google.protobuf.MessageOptions {
+  optional MessageType message_type = 50001;
+  optional Meta meta = 50002;
+}
+
+enum MessageType {
+  MESSAGE_TYPE_UNSPECIFIED = 0;
+  MESSAGE_TYPE_TELEMETRY = 1;
+  MESSAGE_TYPE_TELECOMMAND = 2;
+  MESSAGE_TYPE_TELECONFIG = 3;
+}
+
+message Meta {
+	map<string, string> tags = 1; // tags are used to uniquely identify the device
+}
+
+extend google.protobuf.FieldOptions {
+  optional TimestampType timestamp = 50101;
+  optional FieldMeta field_meta = 50102;
+}
+
+message FieldMeta {
+	map<string, string> tags = 1; // tags are used to uniquely identify the device
+}
+
+enum TimestampType {
+  TIMESTAMP_TYPE_UNSPECIFIED = 0;
+  TIMESTAMP_TYPE_SEC = 1; // Represents seconds of UTC time since Unix epoch (int64)
+  TIMESTAMP_TYPE_MICRO = 2; // Represents microseconds of UTC time since Unix epoch (int64)
+  TIMESTAMP_TYPE_MILLI = 3; // Represents milliseconds of UTC time since Unix epoch (int64)
+  TIMESTAMP_TYPE_NANO = 4; // Represents nanoseconds of UTC time since Unix epoch (int64)
+  TIMESTAMP_TYPE_FRACTION = 5; // Represents seconds of UTC time since Unix epoch (int64) and non-negative fractions of a second at nanosecond resolution (int32
+}
+
+message Timestamp {
+  // Represents seconds of UTC time since Unix epoch
+  // 1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
+  // 9999-12-31T23:59:59Z inclusive.
+  int64 seconds = 1;
+
+  // Non-negative fractions of a second at nanosecond resolution. Negative
+  // second values with fractions must still have non-negative nanos values
+  // that count forward in time. Must be from 0 to 999,999,999
+  // inclusive.
+  int32 nanos = 2;
+}
+
+/*
+    A simple empty message to return as a response
+*/
+message Void {
+
+}
+
+/*
+    An error message to mimic http error messages
+*/
+message Error {
+    string message = 1;  // Human-readable error message.
+}
