+++ examples/tutorials/device/buf.gen.yaml
