+++ examples/tutorials/device/proto/tutorial_device/v1/schema.proto
